+++ packages/connect/CHANGELOG.md
@@ -16,3 +16,7 @@
     -   `CardanoTxWitnessType` => `PROTO.CardanoTxWitnessType`
     -   removed unused `DEVICE.WAIT_FOR_SELECTION`
     -   removed unused `UI.CHANGE_ACCOUNT`
+
+## Added:
+
+-   `getOwnershipId` method